Remove RepeatingReferenceComponent, because it only had a single use and could be moved there
