+++ src/Altinn.App.Core/Models/Layout/Components/RepeatingGroupComponent.cs
@@ -9,7 +9,7 @@ namespace Altinn.App.Core.Models.Layout.Components;
 /// <summary>
 /// Component specialization for repeating groups with maxCount > 1
 /// </summary>
-public sealed class RepeatingGroupComponent : Base.RepeatingReferenceComponent
+public sealed class RepeatingGroupComponent : Base.BaseLayoutComponent
 {
     /// <summary>
     /// Parser for RepeatingGroupComponent
@@ -120,6 +120,99 @@ public sealed class RepeatingGroupComponent : Base.RepeatingReferenceComponent
     /// </summary>
     public required IReadOnlyList<GridComponent.GridRowConfig> RowsAfter { get; init; }
 
+    /// <summary>
+    /// Model binding for the group that defines the number of repetitions of the repeating group.
+    /// </summary>
+    public required ModelBinding GroupModelBinding { get; init; }
+
+    /// <summary>
+    /// The expression that determines if the row is hidden.
+    /// </summary>
+    public required Expression HiddenRow { get; init; }
+
+    /// <summary>
+    /// List of references to child components that are repeated for each row in the repeating group.
+    /// </summary>
+    public required IReadOnlyList<string> RepeatingChildReferences { get; init; }
+
+    /// <summary>
+    /// References to child components that are not repeated and comes before the repeating group
+    /// </summary>
+    public required IReadOnlyList<string> BeforeChildReferences { get; init; }
+
+    /// <summary>
+    /// References to child components that are not repeated and comes after the repeating group
+    /// </summary>
+    public required IReadOnlyList<string> AfterChildReferences { get; init; }
+
+    /// <summary>
+    /// References to the components that are used for the child contexts of this component
+    /// </summary>
+    private Dictionary<string, Base.BaseLayoutComponent>? _claimedChildrenLookup;
+
+    // used for some tests to ensure hierarchy is correct
+    internal IEnumerable<Base.BaseComponent>? AllChildren => _claimedChildrenLookup?.Values;
+
+    /// <inheritdoc />
+    public override void ClaimChildren(
+        Dictionary<string, Base.BaseLayoutComponent> unclaimedComponents,
+        Dictionary<string, string> claimedComponents
+    )
+    {
+        if (
+            GroupModelBinding.Field is null
+            || RepeatingChildReferences is null
+            || BeforeChildReferences is null
+            || AfterChildReferences is null
+        )
+        {
+            throw new UnreachableException(
+                $"{nameof(RepeatingGroupComponent)} must initialize {nameof(GroupModelBinding)}, {nameof(RepeatingChildReferences)}, {nameof(HiddenRow)}, {nameof(BeforeChildReferences)} and {nameof(AfterChildReferences)} in its constructor."
+            );
+        }
+
+        var components = new Dictionary<string, Base.BaseLayoutComponent>();
+        foreach (var componentId in BeforeChildReferences.Concat(RepeatingChildReferences).Concat(AfterChildReferences))
+        {
+            if (unclaimedComponents.Remove(componentId, out var component))
+            {
+                claimedComponents[componentId] = Id;
+            }
+            else
+            {
+                // Invalid reference. Throw the appropriate exception.
+                if (claimedComponents.TryGetValue(componentId, out var claimedComponent))
+                {
+                    throw new ArgumentException(
+                        $"Attempted to claim child with id {componentId} to component {Id}, but it has already been claimed by {claimedComponent}."
+                    );
+                }
+                throw new ArgumentException(
+                    $"Attempted to claim child with id {componentId} to component {Id}, but the componentId does not exist"
+                );
+            }
+
+            if (!components.TryAdd(component.Id, component))
+            {
+                throw new ArgumentException($"Component with id {component.Id} is claimed twice by {Id}.");
+            }
+        }
+
+        _claimedChildrenLookup = components;
+    }
+
+    internal static int[] GetSubRowIndexes(int[]? baseIndexes, int index)
+    {
+        if (baseIndexes is null || baseIndexes.Length == 0)
+        {
+            return new[] { index };
+        }
+        var result = new int[baseIndexes.Length + 1];
+        Array.Copy(baseIndexes, result, baseIndexes.Length);
+        result[^1] = index;
+        return result;
+    }
+
     /// <inheritdoc />
     public override async Task<ComponentContext> GetContext(
         LayoutEvaluatorState state,
@@ -129,7 +222,7 @@ public sealed class RepeatingGroupComponent : Base.RepeatingReferenceComponent
     )
     {
         if (
-            ClaimedChildrenLookup is null
+            _claimedChildrenLookup is null
             || RepeatingChildReferences is null
             || BeforeChildReferences is null
             || GroupModelBinding.Field is null
@@ -198,8 +291,8 @@ public sealed class RepeatingGroupComponent : Base.RepeatingReferenceComponent
         Dictionary<string, LayoutSetComponent> layoutsLookup
     )
     {
-        Debug.Assert(ClaimedChildrenLookup is not null, "Must call ClaimChildren before GetContext");
-        if (!ClaimedChildrenLookup.TryGetValue(componentId, out var childComponent))
+        Debug.Assert(_claimedChildrenLookup is not null, "Must call ClaimChildren before GetContext");
+        if (!_claimedChildrenLookup.TryGetValue(componentId, out var childComponent))
         {
             throw new ArgumentException($"Child component with id {componentId} not found in claimed children.");
         }
@@ -230,7 +323,8 @@ public class RepeatingGroupRowComponent : Base.BaseComponent
         TextResourceBindings = repeatingReferenceComponent.TextResourceBindings;
         if (DataModelBindings.TryGetValue("group", out var groupBinding))
         {
-            Hidden = repeatingReferenceComponent.HiddenRow;
+            // Groups must have a group binding, so this code should never run.
+            throw new UnreachableException("RepeatingGroupComponent must have a group binding.");
         }
         else
             Hidden = new Expression(
